+++ internal/api/version_check.go
@@ -6,6 +6,17 @@ import (
 	"github.com/Masterminds/semver"
 )
 
+// NOTE: A version with a prerelease suffix (e.g. the "-rc.3" of "3.35.1-rc.3") is not
+// considered by semver to satisfy a constraint without a prerelease suffix, regardless of
+// whether or not the major/minor/patch version is greater than or equal to that of the
+// constraint.
+//
+// For example, the version "3.35.1-rc.3" is not considered to satisfy the constraint ">=
+// 3.23.0". This is likely not the expected outcome. However, the same version IS
+// considered to satisfy the constraint "3.23.0-0". Thus, it is recommended to pass a
+// constraint with a minimum prerelease version suffix attached if comparisons to
+// prerelease versions are ever expected. See
+// https://github.com/Masterminds/semver#working-with-prerelease-versions for more.
 func CheckSourcegraphVersion(version, constraint, minDate string) (bool, error) {
 	if version == "dev" || version == "0.0.0+dev" {
 		return true, nil
@@ -26,5 +37,6 @@ func CheckSourcegraphVersion(version, constraint, minDate string) (bool, error) 
 	if err != nil {
 		return false, err
 	}
+
 	return c.Check(v), nil
 }